VarysCode/requirements.txt
numpy~=1.21.2
opencv-python~=3.3.0.9
PyAutoGUI~=0.9.53
mss~=6.1.0
Pillow~=8.3.2
pynput~=1.7.3